refactor: abstract memory and wasi symbols
