+++ runtime/libc/wasi/include/wasi_impl.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <assert.h>
 #include <stdio.h>
 #include <stdlib.h>
